ENHANCEMENT: issue #18 and #4: setup.py
adds a setup.py script which allows to install HPOlib. It is
not yet possible to install runsolver and/or optimizers
automatically as well.

HPOlib-run is, having installed HPOlib, the script run an experiment.

wrapping.py adds the HPOlib directory to the pythonpath to be able to
run experiments without installing the lib.

Other changes solve import problems.