FIX: SMAC parameter intensification percentage=0
This makes SMAC (hopefully) deterministic

If the intensification percentage is larger than 0, then an amount of
time greater than zero seconds is spent on intensification. The amount
of time depends on the time needed to build the model and is therefore
not deterministic.

It can also occur that the number of target algorithm runs in the
intensification timespan differs across runs.

We performed tests on branin and har6 and they resulted in a small
performance gain compared to the values on the NIPS workshop poster.