REFACTOR: #14: Make optimizers more indepentend (Randomtpe)
All tpe related files are now in HPOlib/optimizers/tpe/
There you find the parser, the *Default.cfg. and the .py. The config tells
where to find the optimizer source, which can be everywhere on a system

Calling HPO with randomtpe:

Now we have two optimizers in tpe
 HPOlib-run -o tpe # will fail
We can call
 HPOlib-run -o tpe/h
 HPOlib-run -o tpe/r

---

also added randomtpe to branin