add smac_2_08_master as an optimizer
this version of smac supports model sharing, see manual
http://www.cs.ubc.ca/labs/beta/Projects/SMAC/v2.08.00/manual.pdf

To use this option add shared_model=23 to the config.cfg, where 23 can
be any integer-valued number. All SMAC optimization runs will run in this
directory and the results will be collected in ONE .pkl

The updated version can be used with -o optimizers/smac/smac_2_08_00-master

Note: Using a shared model it is not possible to use a temporary output dir.

Note: With this commit the updated version of smac will be downloaded
automatically during installation.