REFACTOR split AND in if condition into two if conditions to improve speed
The second condition introduced a huge overhead into wrapping.py.
It's a call to process.children(). This call is only necessary
when the child process of wrapping.py (the optimizer) isn't
running any more. By putting the function call to process.children()
inside the if block, process.children() will only be called once
the optimizer finished. This greatly improves performance.